+++ src/gdal_utils.cpp
@@ -16,10 +16,12 @@
 #include "gdal_sf_pkg.h"
 
 // [[Rcpp::export]]
-Rcpp::CharacterVector CPL_gdalinfo(Rcpp::CharacterVector obj, Rcpp::CharacterVector options) {
+Rcpp::CharacterVector CPL_gdalinfo(Rcpp::CharacterVector obj, Rcpp::CharacterVector options, 
+		Rcpp::CharacterVector oo) {
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALInfoOptions* opt = GDALInfoOptionsNew(options_char.data(), NULL);
-	GDALDatasetH ds = GDALOpen((const char *) obj[0], GA_ReadOnly);
+	GDALDatasetH ds = GDALOpenEx((const char *) obj[0], GA_ReadOnly, NULL, oo_char.data(), NULL);
 	if (ds == NULL)
 		return 1; // #nocov
 	char *ret_val = GDALInfo(ds, opt);
@@ -34,13 +36,14 @@ Rcpp::CharacterVector CPL_gdalinfo(Rcpp::CharacterVector obj, Rcpp::CharacterVec
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalwarp(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	std::vector<GDALDatasetH> src_pt(src.size());
 	for (int i = 0; i < src.size(); i++)
-		src_pt[i] = GDALOpen((const char *) src[i], GA_ReadOnly);
+		src_pt[i] = GDALOpenEx((const char *) src[i], GA_ReadOnly, NULL, oo_char.data(), NULL);
 
 	std::vector <char *> options_char = create_options(options, true);
 	GDALWarpAppOptions* opt = GDALWarpAppOptionsNew(options_char.data(), NULL);
@@ -57,13 +60,15 @@ Rcpp::LogicalVector CPL_gdalwarp(Rcpp::CharacterVector src, Rcpp::CharacterVecto
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalrasterize(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options, bool overwrite = false) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo, bool overwrite = false) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALRasterizeOptions* opt =  GDALRasterizeOptionsNew(options_char.data(), NULL);
 
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_VECTOR | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_VECTOR | GA_ReadOnly, 
+		NULL, oo_char.data(), NULL);
 	if (src_pt == NULL)
 		Rcpp::stop("source dataset not found");
 	unset_error_handler();
@@ -88,13 +93,15 @@ Rcpp::LogicalVector CPL_gdalrasterize(Rcpp::CharacterVector src, Rcpp::Character
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdaltranslate(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true);
 	GDALTranslateOptions* opt =  GDALTranslateOptionsNew(options_char.data(), NULL);
 
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, 
+		NULL, oo_char.data(), NULL);
 	if (src_pt == NULL)
 		return 1; // #nocov
 	GDALDatasetH result = GDALTranslate((const char *) dst[0], src_pt, opt, &err);
@@ -108,13 +115,15 @@ Rcpp::LogicalVector CPL_gdaltranslate(Rcpp::CharacterVector src, Rcpp::Character
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalvectortranslate(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALVectorTranslateOptions* opt =  GDALVectorTranslateOptionsNew(options_char.data(), NULL);
 
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_VECTOR | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_VECTOR | GA_ReadOnly, NULL, 
+		oo_char.data(), NULL);
 	if (src_pt == NULL)
 		return 1; // #nocov
 	// GDALDatasetH dst_pt = GDALOpen((const char *) dst[0], GA_Update);
@@ -129,19 +138,28 @@ Rcpp::LogicalVector CPL_gdalvectortranslate(Rcpp::CharacterVector src, Rcpp::Cha
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalbuildvrt(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALBuildVRTOptions* opt = GDALBuildVRTOptionsNew(options_char.data(), NULL);
 
+	/*
 	std::vector<const char *> srcpt(src.size());
 	for (int i = 0; i < src.size(); i++)
 		srcpt[i] = (const char *) src[i];
+	*/
+	std::vector<GDALDatasetH> srcpt(src.size());
+	for (int i = 0; i < src.size(); i++)
+		srcpt[i] = GDALOpenEx((const char *) src[i], GDAL_OF_RASTER | GA_ReadOnly, NULL, 
+			oo_char.data(), NULL);
 
-	GDALDatasetH result = GDALBuildVRT((const char *) dst[0], src.size(), NULL, srcpt.data(), opt, &err);
+	GDALDatasetH result = GDALBuildVRT((const char *) dst[0], src.size(), srcpt.data(), NULL, opt, &err);
 
 	GDALBuildVRTOptionsFree(opt);
+	for (int i = 0; i < src.size(); i++)
+		GDALClose(srcpt[i]);
 	if (result != NULL)
 		GDALClose(result);
 	return result == NULL || err;
@@ -149,13 +167,16 @@ Rcpp::LogicalVector CPL_gdalbuildvrt(Rcpp::CharacterVector src, Rcpp::CharacterV
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdaldemprocessing(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options, Rcpp::CharacterVector processing, Rcpp::CharacterVector colorfilename) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector processing, Rcpp::CharacterVector colorfilename,
+		Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALDEMProcessingOptions* opt =  GDALDEMProcessingOptionsNew(options_char.data(), NULL);
 
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, 
+		NULL, oo_char.data(), NULL);
 	if (src_pt == NULL)
 		Rcpp::stop("cannot open source dataset"); // #nocov
 	GDALDatasetH result = GDALDEMProcessing((const char *) dst[0], src_pt, 
@@ -172,14 +193,15 @@ Rcpp::LogicalVector CPL_gdaldemprocessing(Rcpp::CharacterVector src, Rcpp::Chara
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalnearblack(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALNearblackOptions* opt =  GDALNearblackOptionsNew(options_char.data(), NULL);
 
 	// GDALDatasetH src_pt = GDALOpen((const char *) src[0], GA_ReadOnly);
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_RASTER | GA_ReadOnly, NULL, oo_char.data(), NULL);
 	GDALDatasetH dst_pt = GDALOpen((const char *) dst[0], GA_Update);
 	GDALDatasetH result = GDALNearblack(NULL, dst_pt, src_pt, opt, &err);
 	GDALNearblackOptionsFree(opt);
@@ -194,13 +216,15 @@ Rcpp::LogicalVector CPL_gdalnearblack(Rcpp::CharacterVector src, Rcpp::Character
 
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdalgrid(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 
 	int err = 0;
 	std::vector <char *> options_char = create_options(options, true);
+	std::vector <char *> oo_char = create_options(oo, true); // open options
 	GDALGridOptions* opt =  GDALGridOptionsNew(options_char.data(), NULL);
 
-	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_ALL | GA_ReadOnly, NULL, NULL, NULL);
+	GDALDatasetH src_pt = GDALOpenEx((const char *) src[0], GDAL_OF_ALL | GA_ReadOnly, 
+		NULL, oo_char.data(), NULL);
 	GDALDatasetH result = GDALGrid((const char *) dst[0], src_pt, opt, &err);
 	GDALGridOptionsFree(opt);
 	if (src_pt != NULL)
@@ -213,47 +237,50 @@ Rcpp::LogicalVector CPL_gdalgrid(Rcpp::CharacterVector src, Rcpp::CharacterVecto
 #else
 #include "Rcpp.h"
 
-Rcpp::CharacterVector CPL_gdalinfo(Rcpp::CharacterVector obj, Rcpp::CharacterVector options) {
+Rcpp::CharacterVector CPL_gdalinfo(Rcpp::CharacterVector obj, Rcpp::CharacterVector options, 
+		Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalwarp(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalrasterize(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options, bool overwrite = false) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo, bool overwrite = false) {
+
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdaltranslate(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalvectortranslate(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalbuildvrt(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdaldemprocessing(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options, Rcpp::CharacterVector processing, Rcpp::CharacterVector colorfilename) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector processing, Rcpp::CharacterVector colorfilename,
+		Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalnearblack(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 
 Rcpp::LogicalVector CPL_gdalgrid(Rcpp::CharacterVector src, Rcpp::CharacterVector dst,
-		Rcpp::CharacterVector options) {
+		Rcpp::CharacterVector options, Rcpp::CharacterVector oo) {
 	Rcpp::stop("GDAL version >= 2.1 required for gdal_utils");
 }
 #endif
@@ -262,15 +289,16 @@ Rcpp::LogicalVector CPL_gdalgrid(Rcpp::CharacterVector src, Rcpp::CharacterVecto
 // https://www.gdal.org/warptut.html :
 // [[Rcpp::export]]
 Rcpp::LogicalVector CPL_gdal_warper(Rcpp::CharacterVector infile, Rcpp::CharacterVector outfile,
-		Rcpp::IntegerVector options)
-{
+		Rcpp::IntegerVector options, Rcpp::CharacterVector oo) {
+
+	std::vector <char *> oo_char = create_options(oo, true); // open options
     GDALDatasetH  hSrcDS, hDstDS;
     // Open input and output files.
     GDALAllRegister();
-    hSrcDS = GDALOpen( infile[0], GA_ReadOnly );
+    hSrcDS = GDALOpenEx( infile[0], GA_ReadOnly, NULL, oo_char.data(), NULL );
 	if (hSrcDS == NULL)
 		Rcpp::stop("input file not found");
-    hDstDS = GDALOpen( outfile[0], GA_Update );
+    hDstDS = GDALOpenEx( outfile[0], GA_Update, NULL, NULL, NULL ); // TODO: destination open options?
 	if (hDstDS == NULL)
 		Rcpp::stop("could not open output file for writing");
     // Setup warp options.