+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * reorganize `crs` objects to reflect our post-proj4string world (#1146; #1225)
 
+* `gdal_utils` processes `-oo` open options properly; https://github.com/ITSLeeds/geofabric/issues/12
+
 * `st_sample` directly interfaces `spatstat` sampling methods, e.g. `type = "Thomas"` calls `spatstat::rThomas` after converting input arguments (window) and converts returned `ppp` object to `sf`'s `POINT` geometries; #1204 with help from Ege Rubak and Jakub Nowosad
 
 * `sf_project` gains an option `keep = TRUE` to return `Inf` values for points not projectable; #1228